Update #3
Reworked a lot of php functions, cleaned up so that the queries are now
relying on the ref id in the datagrams table versus title.
Created timestamp column in datagrams table
Made it so that datagrams and displayed in descending order of time it
was created.
Changed the way links are posted so that a time is created as well
Created a create photo datagram menu as well, same as link
Done:
Reworked PHP to print in descending order by time
Create a form for posting photos
Made home page show user's own posts
To do:
Create a form for posting videos
Limit number of datagrams to be displayed at once
Show more datagrams as you scroll down
Setup a messaging system
Friends page?